feat: add shared workflows to help with releases
The first workflow extracts the latest version tag from the project's
change log and outputs it. Note that as some sort of "safety net"
the release date for that version must match the current date.
The idea is that the change log should have been updated recently.

The second workflow can be used to create an annotated git tag.
Again the project's change log is used to create the tag's message.