feat(release): integrate "release-please" action into build workflow
The "release-please" action is run after the images have been built
and tested. This either creates or updates the release pull request
or, in case it was merged, creates a release. After that the above
images are tagged accordingly.

See https://medium.com/@koladilip/embracing-automation-in-versioning-the-power-of-release-please-github-action-4241bd8f3b54
and https://github.com/googleapis/release-please-action