feat(nightwatch): pass cluster to baseproject action
Since adding the ability to deploy to multiple clusters it is no longer
possible to know which is 'the' cluster for an environment. We need to
specify the target cluster manually. With the 1to1 correlation between
env and cluster missing it seems useful to also make it mandatory to
specify the environment, to be explicit.